TINY-10147: Rename BespokeAriaLabelTest to DropdownAriaLabelTest and remove context scopes
